Prepare for final release
Updated README files to include instructions for installation and user
guide.

Changed root package from duke to dac.

Set all font in GUI to Arial to support cross-platform.

Updated JUNIT test files. Removed text-ui-test folder.

Extracted some code into constants for better readability in
DialogBox.java.
