+++ README.md
@@ -1,24 +1,13 @@
-# Duke project template
+# Dog-and-Cat (DaC)
+Dog-and-Cat (DaC) is a chat-bot that keeps track of a list of tasks.
 
-This is a project template for a greenfield Java project. It's named after the Java mascot _Duke_. Given below are instructions on how to use it.
+## Setting up
 
-## Setting up in Intellij
+Prerequisites: JDK 11.
+Download here: https://www.oracle.com/sg/java/technologies/javase/jdk11-archive-downloads.html.
 
-Prerequisites: JDK 11, update Intellij to the most recent version.
+Download the JAR file from: https://github.com/Joel-Sung/ip/releases/tag/A-Release
 
-1. Open Intellij (if you are not in the welcome screen, click `File` > `Close Project` to close the existing project first)
-1. Open the project into Intellij as follows:
-   1. Click `Open`.
-   1. Select the project directory, and click `OK`.
-   1. If there are any further prompts, accept the defaults.
-1. Configure the project to use **JDK 11** (not other versions) as explained in [here](https://www.jetbrains.com/help/idea/sdk.html#set-up-jdk).<br>
-   In the same dialog, set the **Project language level** field to the `SDK default` option.
-3. After that, locate the `src/main/java/Duke.java` file, right-click it, and choose `Run Duke.main()` (if the code editor is showing compile errors, try restarting the IDE). If the setup is correct, you should see something like the below as the output:
-   ```
-   Hello from
-    ____        _        
-   |  _ \ _   _| | _____ 
-   | | | | | | | |/ / _ \
-   | |_| | |_| |   <  __/
-   |____/ \__,_|_|\_\___|
-   ```
+Run the JAR file from any file that does not require administrator access, and you are **DONE**!!!
+
+Refer to the user guide for information on how to use the program: https://github.com/Joel-Sung/ip/tree/master/docs
