--- src/main/java/Duke/exception/DukeException.java
@@ -1,10 +1,0 @@
-package duke.exception;
-
-/**
- * Represents exceptions unique to Duke.
- */
-public class DukeException extends Exception {
-    public DukeException(String str) {
-        super(str);
-    }
-}
