--- src/main/java/Duke/exception/InvalidInputException.java
@@ -1,29 +1,0 @@
-package duke.exception;
-
-/**
- * Represents an invalid input. Thrown when user input is invalid.
- */
-public class InvalidInputException extends DukeException {
-
-    private String err;
-
-    /**
-     * Constructor.
-     *
-     * @param str The invalid input.
-     */
-    public InvalidInputException(String str) {
-        super(str);
-        this.err = str;
-    }
-
-    /**
-     * Returns a String representation of the exception.
-     *
-     * @return String representation of the exception.
-     */
-    @Override
-    public String toString() {
-        return "Invalid input: " + err;
-    }
-}
