+++ src/main/java/dac/exception/DukeException.java
@@ -1,0 +1,10 @@
+package dac.exception;
+
+/**
+ * Represents exceptions unique to Duke.
+ */
+public class DukeException extends Exception {
+    public DukeException(String str) {
+        super(str);
+    }
+}
