+++ src/main/resources/view/DialogBox.fxml
@@ -10,7 +10,7 @@
 
 <fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="490.0" minHeight="100.0" minWidth="490.0" style="-fx-alignment: Bottom_RIGHT;" type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml">
     <children>
-        <Label fx:id="dialog" alignment="CENTER" maxHeight="1.7976931348623157E308" minHeight="75.0" minWidth="370.0" text="Long ass text woooooooooohoooooooooooooooooooooooooooLong ass text woooooooooohoooooooooooooooooooooooooooLong ass text woooooooooohooooooooooooooooooooooooooo" textFill="#000000" wrapText="true">
+        <Label fx:id="dialog" alignment="CENTER" maxHeight="1.7976931348623157E308" minHeight="75.0" minWidth="370.0" style="-fx-font-family: Arial; -fx-font-size: 12;" text="Long ass text woooooooooohoooooooooooooooooooooooooooLong ass text woooooooooohoooooooooooooooooooooooooooLong ass text woooooooooohooooooooooooooooooooooooooo" textFill="#000000" wrapText="true">
          <font>
             <Font name="Comic Sans MS Italic" size="10.0" />
          </font>
