--- text-ui-test/EXPECTED.TXT
@@ -1,149 +1,0 @@
--------------------------------------------------------------------------------------------------------------
-
-	Hello! I'm Duke!
-
-	There are no items in the task list.
-
-	What can I do for you?
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid instruction: wrongcommand is not a valid instruction.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid input: todo task needs a description.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Got it. I've added this task:
-		[T][ ] this
-	Now you have 1 tasks in the list.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid input: deadline task needs a description.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid input: deadline task must have details before and after /by.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Got it. I've added this task:
-		[D][ ] return book (by: Dec 20 2000 12:00PM)
-	Now you have 2 tasks in the list.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Got it. I've added this task:
-		[E][ ] project meeting (at: Aug 22 2021 7:15AM to 9:15AM)
-	Now you have 3 tasks in the list.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Got it. I've added this task:
-		[E][ ] project meeting 2 (at: Dec 20 2000 12:00PM to 12:00AM)
-	Now you have 4 tasks in the list.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Tasks in task list:
-		1. [T][ ] this
-		2. [D][ ] return book (by: Dec 20 2000 12:00PM)
-		3. [E][ ] project meeting (at: Aug 22 2021 7:15AM to 9:15AM)
-		4. [E][ ] project meeting 2 (at: Dec 20 2000 12:00PM to 12:00AM)
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Nice! I've marked this task as done:
-		[T][X] this
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid input: Task number does not exist. Complete failed.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Tasks in task list:
-		1. [T][X] this
-		2. [D][ ] return book (by: Dec 20 2000 12:00PM)
-		3. [E][ ] project meeting (at: Aug 22 2021 7:15AM to 9:15AM)
-		4. [E][ ] project meeting 2 (at: Dec 20 2000 12:00PM to 12:00AM)
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid input: Enter the date and time after "date" in this format: YYYY-MM-DD (24hr time).
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	These tasks are due/happening:
-		[D][ ] return book (by: Dec 20 2000 12:00PM)
-		[E][ ] project meeting 2 (at: Dec 20 2000 12:00PM to 12:00AM)
-
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	These tasks are due/happening:
-		[E][ ] project meeting (at: Aug 22 2021 7:15AM to 9:15AM)
-
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Invalid input: Enter the date and time after "date" in this format: YYYY-MM-DD (24hr time).
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Here are the matching tasks in your list:
-
-	[E][ ] project meeting (at: Aug 22 2021 7:15AM to 9:15AM)
-	[E][ ] project meeting 2 (at: Dec 20 2000 12:00PM to 12:00AM)
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	There are no matching tasks in your list.
-
--------------------------------------------------------------------------------------------------------------
-
--------------------------------------------------------------------------------------------------------------
-
-	Bye. Hope to see you again soon!
-
--------------------------------------------------------------------------------------------------------------
-
