--- text-ui-test/input.txt
@@ -1,19 +1,0 @@
-wrongcommand
-todo
-todo this
-deadline
-deadline /by
-deadline return book /by 2000-12-20 1200
-event project meeting /at 2021-08-22 0715-0915
-event project meeting 2 /at 2000/12/20 1200-0000
-list
-done 1
-done 10
-list
-date asdfhasdf
-date 2000-12-20 1200
-date 2021/08/22 0715
-date 2021-12-03
-find meeting
-find meeting 3
-bye
